+++ app/mailers/waste_carriers_engine/renewal_mailer.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 
 module WasteCarriersEngine
-  class RenewalMailer < ActionMailer::Base
-    helper "waste_carriers_engine/application"
-    helper "waste_carriers_engine/mailer"
-
+  class RenewalMailer < BaseMailer
     def send_renewal_complete_email(registration)
       @registration = registration
 
@@ -12,7 +9,7 @@ module WasteCarriersEngine
       attachments["WasteCarrierRegistrationCertificate-#{registration.reg_identifier}.pdf"] = certificate if certificate
 
       mail(to: @registration.contact_email,
-           from: "#{Rails.configuration.email_service_name} <#{Rails.configuration.email_service_email}>",
+           from: from_email,
            subject: I18n.t(".waste_carriers_engine.renewal_mailer.send_renewal_complete_email.subject",
                            reg_identifier: @registration.reg_identifier))
     end
@@ -26,7 +23,7 @@ module WasteCarriersEngine
                        reg_identifier: @transient_registration.reg_identifier)
 
       mail(to: @transient_registration.contact_email,
-           from: "#{Rails.configuration.email_service_name} <#{Rails.configuration.email_service_email}>",
+           from: from_email,
            subject: subject) do |format|
         format.html { render template }
       end
@@ -42,24 +39,6 @@ module WasteCarriersEngine
       else
         "send_renewal_received_pending_conviction_check_email"
       end
-    end
-
-    # We wrap the generation of the pdf in a rescue block, because though it's
-    # not ideal that the user doesn't get their certificate attached if an error
-    # occurs, we also don't want to block their renewal from completing because
-    # of it
-    def generate_pdf_certificate
-      @presenter = CertificatePresenter.new(@registration, view_context)
-      pdf_generator = GeneratePdfService.new(
-        render_to_string(
-          pdf: "certificate",
-          template: "waste_carriers_engine/pdfs/certificate"
-        )
-      )
-      pdf_generator.pdf
-    rescue StandardError => e
-      Airbrake.notify(e, registration_no: @registration.reg_identifier) if defined?(Airbrake)
-      nil
     end
   end
 end
